feat: update gauge handlers for template/instance model and convert values to integers
- Update gauge handlers to work with new template/instance separation
- Convert all gauge values from float64 to int64 for better UX
- Add database migration to change REAL columns to INTEGER
- Update SQLC configuration to generate int64 types
- Update all templates to display integers without decimals
- Fix form validation to use ParseInt instead of ParseFloat
- Update models and test utilities for integer values
- Remove generated *_templ.go files from version control
- Add documentation about not editing generated files

Addresses GitHub PR feedback for integer-based gauge values
Completes task 6: Update gauge handlers to work with template/instance model

+++ internal/db/db_test.go
@@ -58,7 +58,7 @@ func TestQueries_GaugeValues(t *testing.T) {
 		// Create values (all in the same month to test averaging)
 		now := time.Now().UTC()
 		values := []struct {
-			value float64
+			value int64
 			date  time.Time
 		}{
 			{50, now.AddDate(0, 0, -2)},  // 2 days ago
@@ -77,7 +77,7 @@ func TestQueries_GaugeValues(t *testing.T) {
 		assert.Len(t, history, 1) // All values are in the same month
 
 		// Average of 50, 75, 100 should be 75
-		expectedAverage := (50.0 + 75.0 + 100.0) / 3.0
+		expectedAverage := int64((50 + 75 + 100) / 3)
 		assert.Equal(t, expectedAverage, history[0].AverageValue)
 	})
 }
